requirements.txt
tensorflow==2.18.0
opencv-python==4.11.0.86
numpy==2.0.2
pillow==11.1.0
h5py==3.12.1
keras==3.8.0
matplotlib==3.10.0